Update README.md
This project addresses the use of different machine learning algorithms, including Regression (using methods such as Lasso, Ridge, and Elastic Net), Random Forest, and Neural Networks (LSTM and Bi-LSTM), applied to the prediction of sugarcane agricultural yield. The main objective was to evaluate the accuracy and efficiency of these algorithms in estimating agricultural yield and the impacts of climatic variables on production, providing information to support decision-making in the sugar-energy sector. The methodology includes the use of daily climatic variables such as precipitation, maximum and minimum temperature, soil moisture, and evapotranspiration, as well as derived indices such as the Standardized Precipitation Index (SPI). Data preprocessing steps were performed, including normalization, outlier treatment, and division into training and testing sets. The models were evaluated using metrics such as the coefficient of determination (R²), mean absolute error (MAE), and root mean square error (RMSE). The results show that the Random Forest model performed best overall, surpassing the Lasso, Ridge, and Elastic Net methods and the Neural Networks (LSTM and Bi-LSTM) in terms of accuracy and generalization capacity. However, the R² values ranging from 0.48 to 0.62 for the different sugarcane profiles suggest the need to incorporate additional variables or more complex methods to improve predictive power. The proposed approach provides a complementary tool for crop forecasting in the agricultural sector, contributing to reducing uncertainty and optimizing strategic planning in the sugar-energy industry.